server/requirements.txt
autopep8==2.0.2
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
filelock==3.12.0
Flask==2.3.1
Flask-Cors==3.0.10
fsspec==2023.4.0
fugashi==1.2.1
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
jaconv==0.3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pycodestyle==2.10.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
sacremoses==0.0.53
sentencepiece==0.1.98
six==1.16.0
sympy==1.11.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
unidic-lite==1.0.8
urllib3==1.26.15
Werkzeug==2.3.0
zipp==3.15.0
